Last major set of commit to get a region 2 cosmic track processed :-)
The QwTrackingTreeCombine routines now use the QwDetectorInfo in
nearly all region-2-relevant places (region 3 to follow soon), and...

	cosmic tracks are found!

A memory leak is still causing trouble (segfault while deleting the
histograms after writing them to file), so if you rely on running
qwanalysis with a trackingworker you should probably play safe and
wait with updating your local copy (or test in advance that whatever
functionality you need is not hampered).

Preliminary stats: on the first 10 events it seems to get 6 cosmics
tracks, where some of the other 4 are discarded by being outside of
the search path.  The search path is arbitrarily set right now, so
that could be broadened after looking at a larger sample.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@289 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Tracking/include/QwTrackingTreeLine.h
@@ -54,19 +54,19 @@ class QwTrackingTreeLine {
 
     void Print();
 
-    int GetPositionFirst (double binwidth) {
+    double GetPositionFirst (double binwidth) {
       return 0.5 * (a_beg + a_end) * binwidth;
     }; //! Returns position at the first detector plane
 
-    int GetPositionLast (double binwidth) {
+    double GetPositionLast (double binwidth) {
       return 0.5 * (b_beg + b_end) * binwidth;
     }; //! Returns position at the last detector plane
 
-    int GetResolutionFirst (double binwidth) {
+    double GetResolutionFirst (double binwidth) {
       return (a_beg - a_end) * binwidth;
     }; //! Returns resolution at the first detector plane
 
-    int GetResolutionLast (double binwidth) {
+    double GetResolutionLast (double binwidth) {
       return (b_beg - b_end) * binwidth;
     }; //! Returns resolution at the last detector plane
 
@@ -80,7 +80,7 @@ class QwTrackingTreeLine {
     int a_beg, a_end;		/*!< bin in first layer */
     int b_beg, b_end;		/*!< bin in last layer */
 
-    int   numhits;		/*!< number of hits */
+    int   numhits;		/*!< number of hits on this treeline */
     int   nummiss;		/*!< number of planes without hits */
 
     QwHit *hits[2*TLAYERS];	/*!< hitarray */
